在自定义的woker中，增加self.redis方法，让woker可以使用非em-hiredis客户端操作

+++ lib/sad/config.rb
@@ -21,8 +21,8 @@ module Sad
 				@_redis ||= EM::Hiredis.connect(@_redis_url)
 			end
 
-			def interval=(int)
-				@_interval = int.to_i
+			def interval=(float)
+				@_interval = float.to_f
 			end
 
 			def interval
